requirements.txt
gensim==3.8.1
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
joblib==0.14.1
matplotlib==3.1.2
nltk==3.4.5
numpy==1.18.0
oauth2client==4.1.3
pandas==0.25.3
pandas-profiling==2.4.0
py-readability-metrics==1.4.3
pyLDAvis==2.1.2
scikit-learn==0.22
scipy==1.4.1
Scrapy==1.8.0
scrapy-deltafetch==1.2.1
seaborn==0.9.0
sklearn==0.0
streamlit==0.53.0
xgboost==0.90
